requirements.txt
Django>=1.6
django-appconf
django-ldapdb

getconf
vobject
